Rename adapter and layout names
