+++ app/src/main/res/layout/book_fragment.xml
@@ -6,7 +6,6 @@
     android:layout_height="match_parent"
     tools:context=".ui.main.BookFragment">
 
-
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -90,7 +89,7 @@
                         android:layout_height="wrap_content"
                         android:nestedScrollingEnabled="false"
                         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                        tools:listitem="@layout/chapter_item" />
+                        tools:listitem="@layout/group_item" />
 
                 </androidx.appcompat.widget.LinearLayoutCompat>
